+++ weather-app/.idea/libraries/Gradle__android_arch_core_runtime_1_1_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: android.arch.core:runtime-1.1.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runtime-1.1.0.aar/3133d06f8b228049ec5859c7a9bc689c/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runtime-1.1.0.aar/3133d06f8b228049ec5859c7a9bc689c/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runtime-1.1.0.aar/1b93b83b92ee4bb072d452e86e6fca6b/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runtime-1.1.0.aar/1b93b83b92ee4bb072d452e86e6fca6b/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
