+++ weather-app/.idea/libraries/Gradle__android_arch_lifecycle_livedata_core_1_1_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: android.arch.lifecycle:livedata-core-1.1.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/livedata-core-1.1.0.aar/973b8de391ee595df1ea49025caa26ff/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/livedata-core-1.1.0.aar/973b8de391ee595df1ea49025caa26ff/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/livedata-core-1.1.0.aar/01d75f9e16a00656427f5bb232f37e8f/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/livedata-core-1.1.0.aar/01d75f9e16a00656427f5bb232f37e8f/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
