+++ weather-app/.idea/libraries/Gradle__android_arch_lifecycle_runtime_1_1_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: android.arch.lifecycle:runtime-1.1.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runtime-1.1.0.aar/4b58fe0f18a5aca94c026014e1dbb6f7/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runtime-1.1.0.aar/4b58fe0f18a5aca94c026014e1dbb6f7/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runtime-1.1.0.aar/10c01d04380094553aeb389ea858acc4/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runtime-1.1.0.aar/10c01d04380094553aeb389ea858acc4/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
