+++ weather-app/.idea/libraries/Gradle__android_arch_lifecycle_viewmodel_1_1_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: android.arch.lifecycle:viewmodel-1.1.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/viewmodel-1.1.0.aar/d59479f350e97f7d27e8e2c9fc496be1/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/viewmodel-1.1.0.aar/d59479f350e97f7d27e8e2c9fc496be1/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/viewmodel-1.1.0.aar/2e5c5647449e82c35c930d1ab68bbed4/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/viewmodel-1.1.0.aar/2e5c5647449e82c35c930d1ab68bbed4/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
