+++ weather-app/.idea/libraries/Gradle__com_android_support_animated_vector_drawable_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:animated-vector-drawable-27.1.1">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/animated-vector-drawable-27.1.1.aar/0660b02479b863b087485be91f94167d/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/animated-vector-drawable-27.1.1.aar/0660b02479b863b087485be91f94167d/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/animated-vector-drawable-27.1.1.aar/faffb2b9395e88f6d2e4617f71084767/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/animated-vector-drawable-27.1.1.aar/faffb2b9395e88f6d2e4617f71084767/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
