+++ weather-app/.idea/libraries/Gradle__com_android_support_appcompat_v7_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:appcompat-v7-27.1.1">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/appcompat-v7-27.1.1.aar/f5b835bc3480dc2d245f5724e6e07031/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/appcompat-v7-27.1.1.aar/f5b835bc3480dc2d245f5724e6e07031/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/appcompat-v7-27.1.1.aar/d44020e8816c9c1d93f1cf9f127466cd/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/appcompat-v7-27.1.1.aar/d44020e8816c9c1d93f1cf9f127466cd/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
