+++ weather-app/.idea/libraries/Gradle__com_android_support_design_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:design-27.1.1">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/design-27.1.1.aar/cb590fb9fa8fb63a77290508df49bd2a/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/design-27.1.1.aar/cb590fb9fa8fb63a77290508df49bd2a/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/design-27.1.1.aar/d5906a0aa832107aeca498bbf420d490/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/design-27.1.1.aar/d5906a0aa832107aeca498bbf420d490/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
