+++ weather-app/.idea/libraries/Gradle__com_android_support_recyclerview_v7_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:recyclerview-v7-27.1.1">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/recyclerview-v7-27.1.1.aar/a5fc56db2c80e48a0b7dfb1b29dca545/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/recyclerview-v7-27.1.1.aar/a5fc56db2c80e48a0b7dfb1b29dca545/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/recyclerview-v7-27.1.1.aar/9209538374efa1a2d4be801cf4df579f/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/recyclerview-v7-27.1.1.aar/9209538374efa1a2d4be801cf4df579f/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
