+++ weather-app/.idea/libraries/Gradle__com_android_support_support_core_ui_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:support-core-ui-27.1.1">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-core-ui-27.1.1.aar/1af2ac1db0ee46091769ef0aa567c33d/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-core-ui-27.1.1.aar/1af2ac1db0ee46091769ef0aa567c33d/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-core-ui-27.1.1.aar/318e0d143dd4e8d33666f466e30f412d/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-core-ui-27.1.1.aar/318e0d143dd4e8d33666f466e30f412d/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
