+++ weather-app/.idea/libraries/Gradle__com_android_support_support_core_utils_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:support-core-utils-27.1.1">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-core-utils-27.1.1.aar/96bda3465b55e07aa52ed0375e8f6abe/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-core-utils-27.1.1.aar/96bda3465b55e07aa52ed0375e8f6abe/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-core-utils-27.1.1.aar/e8947fb0bf590b53b747d22f1a101720/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-core-utils-27.1.1.aar/e8947fb0bf590b53b747d22f1a101720/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
