+++ weather-app/.idea/libraries/Gradle__com_android_support_support_fragment_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:support-fragment-27.1.1">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-fragment-27.1.1.aar/76beab5b756b1cc127290adb47122227/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-fragment-27.1.1.aar/76beab5b756b1cc127290adb47122227/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-fragment-27.1.1.aar/c7f2e30f3a92eac82358936925e73d53/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-fragment-27.1.1.aar/c7f2e30f3a92eac82358936925e73d53/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
