+++ weather-app/.idea/libraries/Gradle__com_android_support_support_media_compat_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:support-media-compat-27.1.1">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-media-compat-27.1.1.aar/4008443e024580abe55ecc5c2b0297b3/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-media-compat-27.1.1.aar/4008443e024580abe55ecc5c2b0297b3/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-media-compat-27.1.1.aar/152360b892b55a7640f691c957be59cd/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-media-compat-27.1.1.aar/152360b892b55a7640f691c957be59cd/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
