+++ weather-app/.idea/libraries/Gradle__com_android_support_support_v4_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:support-v4-27.1.1">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-v4-27.1.1.aar/77d7fb5731f24d7b09db01bfc7c4a3b1/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-v4-27.1.1.aar/77d7fb5731f24d7b09db01bfc7c4a3b1/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-v4-27.1.1.aar/d8f5d7a9df6ab081a4f89a3575160ae2/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-v4-27.1.1.aar/d8f5d7a9df6ab081a4f89a3575160ae2/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
