+++ weather-app/.idea/libraries/Gradle__com_android_support_support_vector_drawable_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:support-vector-drawable-27.1.1">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-vector-drawable-27.1.1.aar/0dbaa766059b91cbf7eb61320c3e0eea/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-vector-drawable-27.1.1.aar/0dbaa766059b91cbf7eb61320c3e0eea/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-vector-drawable-27.1.1.aar/03fa1aedffba22d18d33bce582cd78e8/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-vector-drawable-27.1.1.aar/03fa1aedffba22d18d33bce582cd78e8/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
