+++ weather-app/.idea/libraries/Gradle__com_android_support_transition_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:transition-27.1.1">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/transition-27.1.1.aar/779b84d9bc4ca313ebe7dc8c20df4024/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/transition-27.1.1.aar/779b84d9bc4ca313ebe7dc8c20df4024/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/transition-27.1.1.aar/c45969f2c4b9de74cc938fa1eab7abac/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/transition-27.1.1.aar/c45969f2c4b9de74cc938fa1eab7abac/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
