+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/MainActivity.java
@@ -23,13 +23,15 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.rperazzo.weatherapp.WeatherManager.FindResult;
-import com.rperazzo.weatherapp.WeatherManager.WeatherService;
 
 import java.util.ArrayList;
 
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
+
+import com.rperazzo.weatherapp.model.*;
+import com.rperazzo.weatherapp.provider.IWeatherProvider;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -42,7 +44,7 @@ public class MainActivity extends AppCompatActivity {
     private ProgressBar mProgressBar;
     private ListView mList;
     private FindItemAdapter mAdapter;
-    private ArrayList<WeatherManager.City> cities = new ArrayList<>();
+    private ArrayList<City> cities = new ArrayList<>();
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -156,7 +158,7 @@ public class MainActivity extends AppCompatActivity {
 
         onStartLoading();
 
-        WeatherService wService = WeatherManager.getService();
+        IWeatherProvider wService = WeatherManager.getService();
         String units = getTemperatureUnit();
         final Call<FindResult> findCall = wService.find(search, units, WeatherManager.API_KEY);
         findCall.enqueue(new Callback<FindResult>() {
@@ -182,16 +184,16 @@ public class MainActivity extends AppCompatActivity {
         return mSharedPref.getString(TEMPERATURE_UNIT_KEY, "metric");
     }
 
-    public class FindItemAdapter extends ArrayAdapter<WeatherManager.City> {
+    public class FindItemAdapter extends ArrayAdapter<City> {
 
-        public FindItemAdapter(Context context, ArrayList<WeatherManager.City> cities) {
+        public FindItemAdapter(Context context, ArrayList<City> cities) {
             super(context, 0, cities);
         }
 
         @Override
         public View getView(int position, View convertView, ViewGroup parent) {
 
-            final WeatherManager.City city = getItem(position);
+            final City city = getItem(position);
 
             if (convertView == null) {
                 convertView = LayoutInflater.from(getContext())
