+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/model/Clouds.java
@@ -1,0 +1,7 @@
+package com.rperazzo.weatherapp.model;
+
+import java.io.Serializable;
+
+public class Clouds implements Serializable {
+    public int all;
+}
