+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/model/Main.java
@@ -1,0 +1,8 @@
+package com.rperazzo.weatherapp.model;
+
+import java.io.Serializable;
+
+public class Main implements Serializable {
+    public double temp;
+    public double pressure;
+}
