+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/model/Sys.java
@@ -1,0 +1,7 @@
+package com.rperazzo.weatherapp.model;
+
+import java.io.Serializable;
+
+public class Sys implements Serializable {
+    public String country;
+}
