+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/model/Weather.java
@@ -1,0 +1,9 @@
+package com.rperazzo.weatherapp.model;
+
+
+import java.io.Serializable;
+
+public class Weather implements Serializable {
+    public String description;
+    public String icon;
+}
