+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/model/Wind.java
@@ -1,0 +1,8 @@
+package com.rperazzo.weatherapp.model;
+
+import java.io.Serializable;
+
+public class Wind implements Serializable {
+
+    public double speed;
+}
